+++ examples/simple-consumer.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # Consumes lines from a Kafka partition and writes them to STDOUT.
 #
 # You need to define the environment variable KAFKA_BROKERS for this
@@ -14,7 +16,7 @@ require "kafka"
 # with e.g. `$stderr` if you want to see what's happening under the hood.
 logger = Logger.new(StringIO.new)
 
-brokers = ENV.fetch("KAFKA_BROKERS")
+brokers = ENV.fetch("KAFKA_BROKERS").split(",")
 
 # Make sure to create this topic in your Kafka cluster or configure the
 # cluster to auto-create topics.