+++ lib/kafka/fetched_batch.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Kafka
 
   # An ordered sequence of messages fetched from a Kafka partition.
@@ -12,7 +14,7 @@ module Kafka
     attr_reader :highwater_mark_offset
 
     # @return [Array<Kafka::FetchedMessage>]
-    attr_reader :messages
+    attr_accessor :messages
 
     def initialize(topic:, partition:, highwater_mark_offset:, messages:)
       @topic = topic