+++ lib/kafka/heartbeat.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Kafka
   class Heartbeat
     def initialize(group:, interval:)
@@ -6,11 +8,13 @@ module Kafka
       @last_heartbeat = Time.now
     end
 
-    def send_if_necessary
-      if Time.now > @last_heartbeat + @interval
-        @group.heartbeat
-        @last_heartbeat = Time.now
-      end
+    def trigger!
+      @group.heartbeat
+      @last_heartbeat = Time.now
+    end
+
+    def trigger
+      trigger! if Time.now > @last_heartbeat + @interval
     end
   end
 end