+++ lib/kafka/protocol.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Kafka
 
   # The protocol layer of the library.
@@ -18,14 +20,20 @@ module Kafka
     TOPIC_METADATA_API = 3
     OFFSET_COMMIT_API = 8
     OFFSET_FETCH_API = 9
-    GROUP_COORDINATOR_API = 10
+    FIND_COORDINATOR_API = 10
     JOIN_GROUP_API = 11
     HEARTBEAT_API = 12
     LEAVE_GROUP_API = 13
     SYNC_GROUP_API = 14
+    DESCRIBE_GROUPS_API = 15
+    LIST_GROUPS_API = 16
     SASL_HANDSHAKE_API = 17
     API_VERSIONS_API = 18
     CREATE_TOPICS_API = 19
+    DELETE_TOPICS_API = 20
+    DESCRIBE_CONFIGS_API = 32
+    ALTER_CONFIGS_API = 33
+    CREATE_PARTITIONS_API = 37
 
     # A mapping from numeric API keys to symbolic API names.
     APIS = {
@@ -35,7 +43,7 @@ module Kafka
       TOPIC_METADATA_API => :topic_metadata,
       OFFSET_COMMIT_API => :offset_commit,
       OFFSET_FETCH_API => :offset_fetch,
-      GROUP_COORDINATOR_API => :group_coordinator,
+      FIND_COORDINATOR_API => :find_coordinator,
       JOIN_GROUP_API => :join_group,
       HEARTBEAT_API => :heartbeat,
       LEAVE_GROUP_API => :leave_group,
@@ -43,6 +51,9 @@ module Kafka
       SASL_HANDSHAKE_API => :sasl_handshake,
       API_VERSIONS_API => :api_versions,
       CREATE_TOPICS_API => :create_topics,
+      DELETE_TOPICS_API => :delete_topics,
+      DESCRIBE_CONFIGS_API => :describe_configs_api,
+      CREATE_PARTITIONS_API => :create_partitions
     }
 
     # A mapping from numeric error codes to exception classes.
@@ -59,7 +70,7 @@ module Kafka
        9 => ReplicaNotAvailable,
       10 => MessageSizeTooLarge,
       12 => OffsetMetadataTooLarge,
-      15 => GroupCoordinatorNotAvailable,
+      15 => CoordinatorNotAvailable,
       16 => NotCoordinatorForGroup,
       17 => InvalidTopic,
       18 => RecordListTooLarge,
@@ -87,19 +98,43 @@ module Kafka
       42 => InvalidRequest
     }
 
+    # A mapping from int to corresponding resource type in symbol.
+    # https://github.com/apache/kafka/blob/trunk/clients/src/main/java/org/apache/kafka/common/resource/ResourceType.java
+    RESOURCE_TYPE_UNKNOWN = 0
+    RESOURCE_TYPE_ANY = 1
+    RESOURCE_TYPE_TOPIC = 2
+    RESOURCE_TYPE_GROUP = 3
+    RESOURCE_TYPE_CLUSTER = 4
+    RESOURCE_TYPE_TRANSACTIONAL_ID = 5
+    RESOURCE_TYPE_DELEGATION_TOKEN = 6
+    RESOURCE_TYPES = {
+      RESOURCE_TYPE_UNKNOWN => :unknown,
+      RESOURCE_TYPE_ANY => :any,
+      RESOURCE_TYPE_TOPIC => :topic,
+      RESOURCE_TYPE_GROUP => :group,
+      RESOURCE_TYPE_CLUSTER => :cluster,
+      RESOURCE_TYPE_TRANSACTIONAL_ID => :transactional_id,
+      RESOURCE_TYPE_DELEGATION_TOKEN => :delegation_token,
+    }
+
+    # Coordinator types. Since Kafka 0.11.0, there are types of coordinators:
+    # Group and Transaction
+    COORDINATOR_TYPE_GROUP = 0
+    COORDINATOR_TYPE_TRANSACTION = 1
+
     # Handles an error code by either doing nothing (if there was no error) or
     # by raising an appropriate exception.
     #
     # @param error_code Integer
     # @raise [ProtocolError]
     # @return [nil]
-    def self.handle_error(error_code)
+    def self.handle_error(error_code, error_message = nil)
       if error_code == 0
         # No errors, yay!
       elsif error = ERRORS[error_code]
-        raise error
+        raise error, error_message
       else
-        raise UnknownError, "Unknown error with code #{error_code}"
+        raise UnknownError, "Unknown error with code #{error_code} #{error_message}"
       end
     end
 
@@ -121,8 +156,8 @@ require "kafka/protocol/fetch_request"
 require "kafka/protocol/fetch_response"
 require "kafka/protocol/list_offset_request"
 require "kafka/protocol/list_offset_response"
-require "kafka/protocol/group_coordinator_request"
-require "kafka/protocol/group_coordinator_response"
+require "kafka/protocol/find_coordinator_request"
+require "kafka/protocol/find_coordinator_response"
 require "kafka/protocol/join_group_request"
 require "kafka/protocol/join_group_response"
 require "kafka/protocol/sync_group_request"
@@ -141,3 +176,15 @@ require "kafka/protocol/sasl_handshake_request"
 require "kafka/protocol/sasl_handshake_response"
 require "kafka/protocol/create_topics_request"
 require "kafka/protocol/create_topics_response"
+require "kafka/protocol/delete_topics_request"
+require "kafka/protocol/delete_topics_response"
+require "kafka/protocol/describe_configs_request"
+require "kafka/protocol/describe_configs_response"
+require "kafka/protocol/alter_configs_request"
+require "kafka/protocol/alter_configs_response"
+require "kafka/protocol/create_partitions_request"
+require "kafka/protocol/create_partitions_response"
+require "kafka/protocol/list_groups_request"
+require "kafka/protocol/list_groups_response"
+require "kafka/protocol/describe_groups_request"
+require "kafka/protocol/describe_groups_response"