+++ spec/functional/batch_consumer_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 describe "Batch Consumer API", functional: true do
   example "consuming messages using the batch API" do
     num_partitions = 15
@@ -9,7 +11,7 @@ describe "Batch Consumer API", functional: true do
     topic = create_random_topic(num_partitions: 15)
 
     Thread.new do
-      kafka = Kafka.new(seed_brokers: kafka_brokers, client_id: "test")
+      kafka = Kafka.new(kafka_brokers, client_id: "test")
       producer = kafka.producer
 
       messages.each do |i|
@@ -23,7 +25,7 @@ describe "Batch Consumer API", functional: true do
 
     threads = 2.times.map do |thread_id|
       t = Thread.new do
-        kafka = Kafka.new(seed_brokers: kafka_brokers, client_id: "test", logger: logger)
+        kafka = Kafka.new(kafka_brokers, client_id: "test", logger: logger)
         consumer = kafka.consumer(group_id: group_id, offset_retention_time: offset_retention_time)
         consumer.subscribe(topic)
 
@@ -56,5 +58,35 @@ describe "Batch Consumer API", functional: true do
 
     expect(received_messages).to eq messages
     expect(duplicates).to eq Set.new
+  end
+
+  example 'support record headers' do
+    topic = create_random_topic(num_partitions: 1)
+    kafka = Kafka.new(seed_brokers: kafka_brokers, client_id: "test")
+    producer = kafka.producer
+    producer.produce(
+      'hello', topic: topic, headers: { 'TracingID' => 'a1', 'SpanID' => 'b2' }
+    )
+    producer.produce(
+      'hello2', topic: topic, headers: { 'TracingID' => 'c3', 'SpanID' => 'd4' }
+    )
+    producer.deliver_messages
+    consumer = kafka.consumer(group_id: SecureRandom.uuid)
+    consumer.subscribe(topic)
+
+    headers = []
+    consumer.each_batch do |batch|
+      batch.messages.each do |message|
+        headers << message.headers
+      end
+      break
+    end
+
+    expect(headers).to eql(
+      [
+        { 'TracingID' => 'a1', 'SpanID' => 'b2' },
+        { 'TracingID' => 'c3', 'SpanID' => 'd4' }
+      ]
+    )
   end
 end