+++ spec/functional/consumer_group_spec.rb
@@ -1,27 +1,18 @@
+# frozen_string_literal: true
+
 describe "Consumer API", functional: true do
-  let(:num_partitions) { 15 }
-  let!(:topic) { create_random_topic(num_partitions: 3) }
   let(:offset_retention_time) { 30 }
 
   example "consuming messages from the beginning of a topic" do
+    topic = create_random_topic(num_partitions: 1)
     messages = (1..1000).to_a
 
-    Thread.new do
-      kafka = Kafka.new(seed_brokers: kafka_brokers, client_id: "test")
+    begin
+      kafka = Kafka.new(kafka_brokers, client_id: "test")
       producer = kafka.producer
 
       messages.each do |i|
-        producer.produce(i.to_s, topic: topic, partition_key: i.to_s)
-
-        if i % 100 == 0
-          producer.deliver_messages
-          sleep 1
-        end
-      end
-
-      (0...num_partitions).each do |i|
-        # Send a tombstone to each partition.
-        producer.produce(nil, topic: topic, partition: i)
+        producer.produce(i.to_s, topic: topic, partition: 0)
       end
 
       producer.deliver_messages
@@ -29,23 +20,27 @@ describe "Consumer API", functional: true do
 
     group_id = "test#{rand(1000)}"
 
-    threads = 2.times.map do |thread_id|
+    mutex = Mutex.new
+    received_messages = []
+
+    consumers = 2.times.map do
+      kafka = Kafka.new(kafka_brokers, client_id: "test", logger: logger)
+      consumer = kafka.consumer(group_id: group_id, offset_retention_time: offset_retention_time)
+      consumer.subscribe(topic)
+      consumer
+    end
+
+    threads = consumers.map do |consumer|
       t = Thread.new do
-        received_messages = []
-
-        kafka = Kafka.new(seed_brokers: kafka_brokers, client_id: "test", logger: logger)
-        consumer = kafka.consumer(group_id: group_id, offset_retention_time: offset_retention_time)
-        consumer.subscribe(topic)
-
         consumer.each_message do |message|
-          if message.value.nil?
-            consumer.stop
-          else
-            received_messages << Integer(message.value)
+          mutex.synchronize do
+            received_messages << message
+
+            if received_messages.count == messages.count
+              consumers.each(&:stop)
+            end
           end
         end
-
-        received_messages
       end
 
       t.abort_on_exception = true
@@ -53,50 +48,32 @@ describe "Consumer API", functional: true do
       t
     end
 
-    received_messages = threads.map(&:value).flatten
+    threads.each(&:join)
 
-    expect(received_messages.sort).to match_array messages
+    expect(received_messages.map(&:value).map(&:to_i)).to match_array messages
   end
 
-  example "consuming messages from the end of a topic" do
-    sent_messages = 1_000
-
-    num_partitions = 1
+  example "consuming messages from a topic that's being written to" do
+    num_partitions = 3
     topic = create_random_topic(num_partitions: num_partitions)
-    group_id = "test#{rand(1000)}"
+    messages = (1..100).to_a
 
-    consumer_thread = Thread.new do
-      received_messages = 0
-
-      kafka = Kafka.new(seed_brokers: kafka_brokers, client_id: "test", logger: logger)
-      consumer = kafka.consumer(group_id: group_id, offset_retention_time: offset_retention_time)
-      consumer.subscribe(topic, start_from_beginning: false)
-
-      consumer.each_message do |message|
-        if message.value.nil?
-          consumer.stop
-        else
-          received_messages += 1
-        end
-      end
-
-      received_messages
-    end
-
-    consumer_thread.abort_on_exception = true
-
-    sleep 1
+    mutex = Mutex.new
+    var = ConditionVariable.new
 
     Thread.new do
-      kafka = Kafka.new(seed_brokers: kafka_brokers, client_id: "test")
+      kafka = Kafka.new(kafka_brokers, client_id: "test")
       producer = kafka.producer
 
-      1.upto(sent_messages) do |i|
-        producer.produce("hello", topic: topic, partition_key: i.to_s)
+      messages.each do |i|
+        producer.produce(i.to_s, topic: topic, partition: i % 3)
 
         if i % 100 == 0
           producer.deliver_messages
-          sleep 1
+
+          mutex.synchronize do
+            var.wait(mutex)
+          end
         end
       end
 
@@ -108,9 +85,35 @@ describe "Consumer API", functional: true do
       producer.deliver_messages
     end
 
-    received_messages = consumer_thread.value
+    group_id = "test#{rand(1000)}"
+    received_messages = []
 
-    expect(received_messages).to eq sent_messages
+    threads = 2.times.map do |thread_id|
+      t = Thread.new do
+        kafka = Kafka.new(kafka_brokers, client_id: "test", logger: logger)
+        consumer = kafka.consumer(group_id: group_id, offset_retention_time: offset_retention_time)
+        consumer.subscribe(topic)
+
+        consumer.each_message do |message|
+          if message.value.nil?
+            consumer.stop
+          else
+            mutex.synchronize do
+              received_messages << Integer(message.value)
+              var.signal
+            end
+          end
+        end
+      end
+
+      t.abort_on_exception = true
+
+      t
+    end
+
+    threads.each(&:join)
+
+    expect(received_messages).to match_array messages
   end
 
   example "stopping and restarting a consumer group" do
@@ -146,5 +149,89 @@ describe "Consumer API", functional: true do
     end
 
     expect(processed_messages).to eq(num_messages * 2)
+  end
+
+  example "consumers process all messages in-order and non-duplicated" do
+    topic = create_random_topic(num_partitions: 2)
+    message_count = 500
+    messages_set_1 = (1..500).to_a
+    messages_set_2 = (501..1000).to_a
+
+    begin
+      kafka = Kafka.new(seed_brokers: kafka_brokers, client_id: "test")
+      producer = kafka.producer
+      messages_set_1.each do |i|
+        producer.produce(i.to_s, topic: topic, partition: 0)
+      end
+      messages_set_2.each do |i|
+        producer.produce(i.to_s, topic: topic, partition: 1)
+      end
+      producer.deliver_messages
+    end
+
+    group_id = "test#{rand(1000)}"
+    received_messages = {}
+
+    consumers = 2.times.map do
+      kafka = Kafka.new(seed_brokers: kafka_brokers, client_id: "test", logger: logger)
+      consumer = kafka.consumer(group_id: group_id)
+      consumer.subscribe(topic)
+      consumer
+    end
+
+    threads = consumers.map do |consumer|
+      t = Thread.new do
+        received_messages[Thread.current] = []
+        puts received_messages
+        consumer.each_message do |message|
+          received_messages[Thread.current] << message
+
+          if received_messages[Thread.current].count == message_count
+            consumer.stop
+          end
+        end
+      end
+      t.abort_on_exception = true
+      t
+    end
+
+    threads.each(&:join)
+
+    received_messages.each do |_thread, messages|
+      values = messages.map(&:value).map(&:to_i)
+      if messages.first.partition == 0
+        expect(values).to eql(messages_set_1)
+      else
+        expect(values).to eql(messages_set_2)
+      end
+    end
+  end
+
+  example 'support record headers' do
+    topic = create_random_topic(num_partitions: 1)
+    kafka = Kafka.new(seed_brokers: kafka_brokers, client_id: "test")
+    producer = kafka.producer
+    producer.produce(
+      'hello', topic: topic, headers: { 'TracingID' => 'a1', 'SpanID' => 'b2' }
+    )
+    producer.produce(
+      'hello2', topic: topic, headers: { 'TracingID' => 'c3', 'SpanID' => 'd4' }
+    )
+    producer.deliver_messages
+    consumer = kafka.consumer(group_id: SecureRandom.uuid)
+    consumer.subscribe(topic)
+
+    headers = []
+    consumer.each_message do |message|
+      headers << message.headers
+      break if headers.length == 2
+    end
+
+    expect(headers).to eql(
+      [
+        { 'TracingID' => 'a1', 'SpanID' => 'b2' },
+        { 'TracingID' => 'c3', 'SpanID' => 'd4' }
+      ]
+    )
   end
 end